add aplit, trim and more

+++ ft_memchr.c
@@ -6,14 +6,14 @@
 /*   By: aljbari <aljbari@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/01 21:57:16 by aljbari           #+#    #+#             */
-/*   Updated: 2024/11/01 22:04:27 by aljbari          ###   ########.fr       */
+/*   Updated: 2024/11/04 18:26:30 by aljbari          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include <stdlib.h>
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int	i;
+	size_t	i;
 	char	*tempP;
 
 	i = 0;
